feat(auth): restructure authentication services and implement remote data source for auth
